chore: move Panel to new components dir
