+++ src/components/panel/Panel.test.js
