+++ src/components/panel/index.ts
