+++ src/index.ts
@@ -1,3 +1,5 @@
-export { Box, Flex, Select, Panel } from './primitives'
+export { Box, Flex, Grid, Select } from './primitives'
+
+export { Panel } from './components'
 
 export { css, styled } from './stitches'
