+++ src/primitives/index.ts
@@ -1,5 +1,4 @@
 export { Box } from './box'
-export { Panel } from './panel'
 export { Flex } from './flex'
 export { Grid } from './grid'
 export { Select } from './select'
